+++ src/WindowsSentinel.Service/Program.cs
@@ -159,7 +159,6 @@ namespace WindowsSentinel.Service
                     services.AddHostedService<ConsultantSignalIngestor>();
                     services.AddHostedService<RegistryMonitor>();
                     services.AddHostedService<GhostProcessMonitor>();
-                    services.AddHostedService<ShellWatchdog>();
                     services.AddHostedService<CriticalServiceGuard>();
                 });
     }